refactor: move functions to helper file
