+++ libs/ledger-live-common/src/wallet-api/helpers.ts
@@ -141,3 +141,20 @@ export function matchCurrencies(
   }
   return matchedCurrencies;
 }
+
+// Copied from https://www.npmjs.com/package/ethereumjs-util
+export const isHexPrefixed = (str: string): boolean => {
+  if (typeof str !== "string") {
+    throw new Error(`[isHexPrefixed] input must be type 'string', received type ${typeof str}`);
+  }
+
+  return str[0] === "0" && str[1] === "x";
+};
+
+// Copied from https://www.npmjs.com/package/ethereumjs-util
+export const stripHexPrefix = (str: string): string => {
+  if (typeof str !== "string")
+    throw new Error(`[stripHexPrefix] input must be type 'string', received ${typeof str}`);
+
+  return isHexPrefixed(str) ? str.slice(2) : str;
+};
